move RenderRoute into its own component and add test for calculateBounds
